TimeLogger and toolset updates
